rename namespace resource to tcnamespace
