+++ service/controller/clusterapi/v19/cluster_resource_set.go
@@ -30,8 +30,8 @@ import (
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/configmap"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/encryptionkey"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/kubeconfig"
-	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/namespace"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/operatorversions"
+	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/tcnamespace"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/tenantclients"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/tiller"
 	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v19/resources/workercount"
@@ -287,18 +287,18 @@ func NewClusterResourceSet(config ClusterResourceSetConfig) (*controller.Resourc
 		}
 	}
 
-	var namespaceResource controller.Resource
+	var tcNamespaceResource controller.Resource
 	{
-		c := namespace.Config{
+		c := tcnamespace.Config{
 			Logger: config.Logger,
 		}
 
-		ops, err := namespace.New(c)
+		ops, err := tcnamespace.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
 
-		namespaceResource, err = toCRUDResource(config.Logger, ops)
+		tcNamespaceResource, err = toCRUDResource(config.Logger, ops)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -374,7 +374,7 @@ func NewClusterResourceSet(config ClusterResourceSetConfig) (*controller.Resourc
 
 		// Following resources manage resources in tenant clusters so they
 		// should be executed last.
-		namespaceResource,
+		tcNamespaceResource,
 		tillerResource,
 		chartOperatorResource,
 		configMapResource,
