+++ service/controller/clusterapi/v19/resources/tcnamespace/desired.go
@@ -1,4 +1,4 @@
-package namespace
+package tcnamespace
 
 import (
 	"context"
